fix: review apply
rename: export name modify

fix: function keyword, div tag remove

fix: convert CSF 3.0

fix: test label

fix: local variable position

fix: reset mock window.print

fix: apply userEvent.setUp

fix: change method name from test to it

fix: mock reset

fix: aria-labelledby test

fire: Certificate/index.tsx

rename: file name temp 변경

rename: Certificate로 변경

fix: Certificate import 수정

fix: tag, useParams query

fix: change to test
