+++ src/main.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import Certificate from "./components/Certificate/certificate";
+import Certificate from "./components/Certificate/Certificate";
 import ErrorPage from "./components/ErrorPage/ErrorPage";
 import Leaderboard from "./components/leaderboard/leaderboard";
 import Progress from "./components/progress/progress";
